# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ticketing application with restful API powered by gin",
        "title": "CHUBA TICKITZ",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "paths": {
...
        "/movies": {
            "get": {
                "description": "Get popular movies, filter movies by title or genres",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "filter movies by genres, title and pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "opsional query for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "opsional query for search title",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "opsional query for filter genres",
                        "name": "genres",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DetailsMovieResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/movies/popular": {
            "get": {
                "description": "Get popular movies, filter movies already rated on every transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Get popular movies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "opsional query for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MoviesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/movies/upcoming": {
            "get": {
                "description": "Get upcoming movies, filter movies that not aired yet",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Get upciming movies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "opsional query for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MoviesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/movies/{movie_id}": {
            "get": {
                "description": "Get details movies, get data by known an id movie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Get details from a movie",
                "parameters": [
                    {
                        "type": "string",
                        "description": "get detail movie by id movie",
                        "name": "movie_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/schedules/{movieid}": {
            "get": {
                "security": [
                    {
                        "JWTtoken": []
                    }
                ],
                "description": "Get schedules movie, for a movie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Get schedule for a movie",
                "parameters": [
                    {
                        "type": "string",
                        "description": "get schedule by this id movie",
                        "name": "movieid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/seats/{schedule_id}": {
            "get": {
                "security": [
                    {
                        "JWTtoken": []
                    }
                ],
                "description": "Get seat that booked  from a movie get by schedule id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seats"
                ],
                "summary": "Get Booked seat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "get booked seat by this schedule id",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SeatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Actor": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.BadRequestResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Example bad request error..."
                },
                "is_success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.BookedSeatBySchedule": {
            "type": "object",
            "properties": {
                "paid_at": {
                    "type": "string"
                },
                "schedule": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "seat_code": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.DetailsMovieResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/models.MovieDetails"
                },
                "is_success": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Example message success..."
                },
                "page": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.Director": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Genre": {
            "type": "object",
            "properties": {
                "genre_id": {
                    "type": "integer"
                },
                "genre_name": {
                    "type": "string"
                }
            }
        },
...
        "models.InternalErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "error": {
                    "type": "string",
                    "example": "Example Internal server error..."
                },
                "is_success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "models.MovieDetails": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Actor"
                    }
                },
                "backdrop_path": {
                    "type": "string"
                },
                "director": {
                    "$ref": "#/definitions/models.Director"
                },
                "director_id": {
                    "type": "integer"
                },
                "duration": {
                    "type": "integer"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Genre"
                    }
                },
                "movie_id": {
                    "type": "integer"
                },
                "overview": {
                    "type": "string"
                },
                "poster_path": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.MovieList": {
            "type": "object",
            "properties": {
                "backdrop_path": {
                    "type": "string"
                },
                "director_id": {
                    "type": "integer"
                },
                "duration": {
                    "type": "integer"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Genre"
                    }
                },
                "movie_id": {
                    "type": "integer"
                },
                "overview": {
                    "type": "string"
                },
                "poster_path": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.MoviesResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MovieList"
                    }
                },
                "is_success": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Example message success..."
                },
                "page": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "models.Schedule": {
            "type": "object",
            "properties": {
                "movie_id": {
                    "type": "integer"
                },
                "schedule": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.ScheduleResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Schedule"
                    }
                },
                "is_success": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Example message success..."
                },
                "page": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "models.SeatResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BookedSeatBySchedule"
                    }
                },
                "is_success": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Example message success..."
                },
                "page": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWTtoken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}